feat(lambda): move interfaces #59
+ Move context and callable interfaces into self-titled folders.
