+++ Domain/Interfaces/Callables/CallableInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace Marsskom\Generator\Domain\Interfaces;
+namespace Marsskom\Generator\Domain\Interfaces\Callables;
 
 interface CallableInterface
 {
