+++ Domain/Interfaces/Callables/PipelineInterface.php
@@ -4,8 +4,6 @@ declare(strict_types = 1);
 
 namespace Marsskom\Generator\Domain\Interfaces\Callables;
 
-use Marsskom\Generator\Domain\Interfaces\CallableInterface;
-
 interface PipelineInterface extends CallableInterface
 {
 }
